Rename 'group' as 'party'
This will keep SQLite happier, as it doesn't quote columns, and GROUP is a
reserved word. In theory we shouldn't care too much about things like
this in the Page classes, and let the scraper do the remappings, but we
don't have a good approach to that yet, so it's simpler to just rename
for now.